Fix test failures and type mismatches
Changes:
- Add Equatable conformance to ConfigDiscoveryError for test expectations
- Re-add findSwiftLintConfig helper function (lost in merge)
- Fix Int32/Int type mismatches in SwiftFormatSmart and SwiftLintCustomSmart
- ProcessRunner.run returns Int32, updated helper function signatures to match

Tests: All 21 tests passing ✓

+++ Sources/SwiftFormatSmart/main.swift
@@ -97,7 +97,7 @@ struct SwiftFormatSmart: ParsableCommand {
     }
 
     /// Throw an error when formatting fails
-    private func throwFormattingError(exitCode: Int, target: String) throws -> Never {
+    private func throwFormattingError(exitCode: Int32, target: String) throws -> Never {
         let errorMsg = ErrorFormatter.format(
             tool: "SwiftFormatSmart",
             errorType: "ExecutionFailed",
@@ -106,6 +106,6 @@ struct SwiftFormatSmart: ParsableCommand {
             fix: "Review SwiftFormat output above for syntax errors or formatting issues",
         )
         Console.error(errorMsg)
-        throw ExitCode(Int32(exitCode))
+        throw ExitCode(exitCode)
     }
 }
